Fix bug

+++ app/src/main/java/com/zen/cendakala/ui/components/CardSurveyHistory.kt
@@ -9,7 +9,6 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.Divider
@@ -18,21 +17,18 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import com.zen.cendakala.R
 import com.zen.cendakala.ui.theme.Black2
 import com.zen.cendakala.ui.theme.Color1
 import com.zen.cendakala.ui.theme.White
-import com.zen.cendakala.utils.Constants
 
 @Composable
 fun CardSurveyHistory(
     modifier: Modifier = Modifier,
     title: String,
-    reward: Long,
+    reward: Int,
 ) {
     Card(
         border = BorderStroke(2.dp, Color1),
@@ -77,7 +73,7 @@ fun CardSurveyHistory(
                 )
                 {
                     Column {
-                        Row{
+                        Row {
                             Image(
                                 painter = painterResource(id = R.drawable.ic_money),
                                 contentDescription = "reward",
